+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Features:
 
-- Support Arch and CoreOS Linux
+- Support Arch, CoreOS, and Gentoo Linux
 - Support FreeBSD, NetBSD, and OpenBSD
 - Support OS X